requirements.txt
asgiref==3.10.0
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
